+++ tests/Pest.php
@@ -1,5 +1,7 @@
 <?php
 
+use Lalalili\SubscriptionCore\Tests\FeatureTestCase;
 use Lalalili\SubscriptionCore\Tests\TestCase;
 
 uses(TestCase::class)->in('Unit');
+uses(FeatureTestCase::class)->in('Feature');